Refactor secrets to auth api pattern
