+++ sales-api/src/config/db/mongoDBConfig.js
@@ -1,5 +1,5 @@
 import mongoose from "mongoose";
-import {MONGO_DB_URL} from "../secrets/secrets.js";
+import {MONGO_DB_URL} from "../constants/secrets.js";
 
 export function connect() {
     mongoose.connect(MONGO_DB_URL, {
