--- sales-api/src/config/secrets/secrets.js
@@ -1,4 +1,0 @@
-const env = process.env;
-
-export const MONGO_DB_URL = env.MONGO_DB_URL ? env.MONGO_DB_URL : "mongodb://admin:123456@localhost:27017";
-export const apiSecret = env.API_SECRET ? env.API_SECRET : 'c2VjcmV0LWFwaS0xMjM0NTYtcGFzc3dvcmQtZW5jb2Rlci02NC1AQEA=';
